Created .csv Writer
Made changes to the PayrecordWriter and the Main Methond in the Program class to execute creating the excel file to display the net, tax and gross along with the EmployeeId.

Also made slight changes to the console display regarding with the displaying of the Net, Gross and Tax properties as it was showing more than 2 significant numbers, as such the function Math.Round(data, 2) has been used. The '2' indicating to what significant number it needs to round up to.

I still need to create NUnit Testing code, the xml documentation and complete the software testing document to completely the assessment.