Slight changes to GetDetails
Has removed the Math.Round to
Property:C and C stands for currency which automatically change the numbers into the money and does the all necessary maths required.

Also the PayRecord code initially used has been commented out and replaced by empty string. As for code that is commented out has been copied to the GetDetails into ResidentPayRecord and WorkingHolidayPayrecord.

I had fixed the issue with the naming convention regarding the exportPath string.