Finalised the internal XML
Going through the document and writing an commented XML description with the summary, remarks and explaining the data type of the parameters to generate the XML document that is done automatically by going to the project settings and properties. And clicking a option inside the build under the Output session.