Final Changes
The testing has been completed. Fixes done:
Logical Errors with TestTax and TestNet were fixed by adding a new variable in the TaxCalculator.cs as well as going into the Work Holiday pay record and changing the YearToDate getter and setter method to be normal and adding new integer variable 'YTD' to display the correct amount to the console.